+++ Utilities/Constants.cpp
@@ -2,7 +2,7 @@
 #include <ctime>
 #include <random>
 
-std::default_random_engine reng(time(nullptr));
+std::default_random_engine reng((int)time(nullptr));
 std::uniform_real_distribution<double> uni_dist(0.0f, 1.0f);
 
 double rand_float()
